Replaced PSR HTTP-message with own, more limited version
We don't want to implement everything in the PSR standard as that is for a full HTTP request cycle, with HTML in mind. This is aimed at APIs and so only implements what is needed to help with speed. Thus we have created our own interfaces. All tests have been updated. Middleware has been implemented. There are still a few missing bits, such as 404 and error responses.